refator: mvc pattern
